Refactor using FormHeader component in forms for flags, segments, and rules
