+++ src/components/segments/SegmentItem.tsx
@@ -2,11 +2,6 @@ import { useState } from "react";
 import { Segment } from "@/types";
 import SegmentsRules from "./SegmentsRules";
 import Modal from "../Modal";
-import FormHeader from "../FormHeader";
-import { useDeleteSegmentMutation } from "@/hooks/segments";
-import toast from "react-hot-toast";
-import ToastTUI from "../ToastTUI";
-import { AxiosError } from "axios";
 import UpdateSegmentForm from "./UpdateSegmentForm";
 
 type SegmentItemProps = Segment;
@@ -14,24 +9,6 @@ type SegmentItemProps = Segment;
 function SegmentItem(props: SegmentItemProps) {
   const [openEdit, setOpenEdit] = useState(false);
   const [openRulesModal, setOpenRulesModal] = useState(false);
-  const { mutateAsync: deleteSegmentMutate } = useDeleteSegmentMutation();
-  const handleDelete = async () => {
-    try {
-      await deleteSegmentMutate(props.sKey);
-      toast.custom(
-        <ToastTUI
-          type="success"
-          message="Segment deleted from the database."
-        />,
-      );
-    } catch (err: unknown) {
-      if (err instanceof AxiosError) {
-        const responseError = err.response?.data.error;
-        toast.custom(<ToastTUI type="error" message={responseError} />);
-      }
-    }
-    setOpenEdit(false);
-  };
 
   return (
     <>
@@ -71,21 +48,11 @@ function SegmentItem(props: SegmentItemProps) {
         </td>
       </tr>
       <Modal open={openEdit} setOpen={setOpenEdit}>
-        <FormHeader
-          resource="segment"
-          onDelete={handleDelete}
-          isDelete={true}
-          directions={`Update ${props.title}.`}
-          action={`Edit segment: ${props.title}`}
-        />
         <UpdateSegmentForm {...props} setOpen={setOpenEdit} />
       </Modal>
       <Modal open={openRulesModal} setOpen={setOpenRulesModal}>
-        <FormHeader
-          action={`Edit rules for ${props.title}`}
-          directions={`Click edit to modify a rule or create a new rule.`}
-        />
         <SegmentsRules
+          title={props.title}
           setOpen={setOpenRulesModal}
           sKey={props.sKey}
           rules={props.rules}
