+++ src/pages/FlagsTable.tsx
@@ -4,7 +4,6 @@ import Button from "@/components/ui/Button";
 import { useFlags } from "@/hooks/flags";
 import Modal from "@/components/Modal";
 import CreateFlagForm from "@/components/flags/CreateFlagForm";
-import FormHeader from "@/components/FormHeader";
 
 export default function FlagsTable() {
   const [openCreateFlagModal, setOpenCreateFlagModal] = useState(false);
@@ -81,10 +80,6 @@ export default function FlagsTable() {
         </div>
       </div>
       <Modal open={openCreateFlagModal} setOpen={setOpenCreateFlagModal}>
-        <FormHeader
-          directions="Define your flag. Note, the flag key will be used to identify your flag in the database and can not be updated once it's created."
-          action="Create a Flag"
-        />
         <CreateFlagForm setOpen={setOpenCreateFlagModal} />
       </Modal>
     </>
